update models and views

+++ invoice1_project/Models/Invoice.cs
@@ -5,9 +5,9 @@ namespace invoice_project.Models
     public class Invoice
     {
         public int Id { get; set; }
-        public int totalCost { get; set; }
+        public float totalCost { get; set; }
         public float tax { get; set; }
-        public int discount { get; set; }
+        public float discount { get; set; }
         public DateTime invoice_date { get; set; }
         [ForeignKey("customer")]
         public int customer_id { get; set; }
@@ -15,7 +15,6 @@ namespace invoice_project.Models
         public int store_id { get; set; }
         public virtual Customer? customer { get; set; }
         public virtual Store? Store { get; set; }
-        //public virtual List<Item>? Items { get; set; }
         public Invoice()
         {
             invoice_date=DateTime.Now;
