+++ invoice1_project/Models/Item.cs
@@ -4,9 +4,9 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int price { get; set; }
+        public float price { get; set; }
         
         public string unit { get; set; }
-        //public virtual List<Invoice>? Invoices { get; set; }
+        
     }
 }
